Move the Kollision icon to the Kollision app dir. (That was post-4.0, so nobody will miss it.)
I'm wondering though how this one is related to the
existing hicolor Kollision icon that already exists
in that directory, without svgz. Seems to be the same
icon, only with different colors - judging from the
current blue/red theme, it seems that the newly moved
("Oxygen") icon is the more current one?

If so, the green/purple hicolor icon should probably
be replaced by the red/blue Oxygen icon, and the latter
simply disappear. In any case, we probably just want
one of these icons, and the associated svgz as well.

CCMAIL: paolo.capriotti@gmail.com

svn path=/trunk/KDE/kdegames/kollision/; revision=807241